fix(configuration): make writer_provider configurable
Under previous structure, LLMs weren't fully configurable because there
was a mix of reading from the `config` object passed in and the defaults
of the `Configuration` object.

`writer_model` was being initialized at the top of `graph.py` before a
configuration was created. It was using defaults from the Configuration
class, rather than takinga config object. This makes it so
`writer_model` is initialized where needed (probably could be factored
out). In addition, it fixes a similar instance of a bug with
planner_model.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>